+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 7.0.0
+
+- Update `filepond` peer dependency to match version 4.0.0
+- Add example folder
+
+
 ## 6.0.1
 
 - Fix typo in README